requirements.txt
semantic-kernel
pydantic
chainlit
python-dotenv
aiosqlite
aiohttp
pandas